HeatControlAPI/API/requirements.txt
aiomysql
fastapi==0.95.2
slowapi==0.1.9
uvicorn[standard]==0.23.1
sqlalchemy==2.0.18
databases[asyncpg]==0.9.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
pandas==2.1.3
pydantic[email]==1.10.7
influxdb-client==1.35.0
python-multipart==0.0.6
requests==2.31.0
mysql-connector-python
timezonefinder==6.2.0